+++ changelog.md
@@ -2,6 +2,7 @@
 - Uniform and Define values on hover code.
 - Fixed GL uniforms reseting on compile.
 - iDate uniform (year, month, day, seconds).
+- User Collections to organize Shaders.
 - Like shaders from Explore page.
 
 # 0.16.1